mpfs-applications: add SMP example that executes from Scratchpad
To demonstrate an SMP application executing from Scratchpad memory, we
add the PI example adopted for two cores, u54_1 and u54_2. We use a
devicetree overlay to specify the area of memory to execute from.

We provide a configuration file with the application which allows the
HSS payload generator to generate a payload that can be written to the
eMMC on the PolarFire SoC Icicle kit.

Signed-off-by: Conor Paxton <conor.paxton@microchip.com>